+++ packages/pages/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@yext/pages",
-  "version": "1.0.0-beta.15",
+  "version": "1.0.0-alpha.281",
   "description": "The default React development toolchain provided by Yext",
   "author": "sumo@yext.com",
   "license": "BSD-3-Clause",
